refactor fight method, refactor class charakter(inheritence) created new modul map created class Maproom + init
