+++ game/map.py
@@ -1,0 +1,8 @@
+class Maproom:
+    Worldmap = []
+    WorldmapCoors = {}
+
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+
