--- komentare.txt
@@ -1,1 +1,0 @@
-Komentáře
